+++ src/main/java/com/globant/itba/storm/bigdatatp2/metricbolts/BoxListFrequencyBolt.java
@@ -44,7 +44,7 @@ public class BoxListFrequencyBolt extends AbstractFrequencyBolt {
     	if (lastSeenValues.containsKey(boxId) && recalculateOnChannelChange) {
     		List<String> prevValues  = lastSeenValues.get(boxId);
     		for (String key : prevValues) {
-    			long prevCount = frequencyTable.get(key);
+    			int prevCount = frequencyTable.get(key);
     			frequencyTable.put(key, prevCount - 1);
     		}
     	}
